Move EventArgs to namespace
