+++ osu.Framework.Tests/Visual/TestCaseInputManager.cs
@@ -9,6 +9,7 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Input;
+using osu.Framework.Input.EventArgs;
 using osu.Framework.Input.Handlers.Mouse;
 using osu.Framework.Input.States;
 using osu.Framework.Testing;
