+++ osu.Framework.Tests/Visual/TestCaseInputQueueChange.cs
@@ -7,7 +7,7 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
-using osu.Framework.Input;
+using osu.Framework.Input.EventArgs;
 using osu.Framework.Input.States;
 using osu.Framework.Testing;
 using OpenTK;
