+++ osu.Framework.Tests/Visual/TestCaseSliderbar.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using osu.Framework.Configuration;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.UserInterface;
-using osu.Framework.Input;
+using osu.Framework.Input.EventArgs;
 using osu.Framework.Input.States;
 using osu.Framework.Testing;
 using OpenTK;
