+++ osu.Framework/Graphics/Containers/ScrollContainer.cs
@@ -3,11 +3,11 @@
 
 using System;
 using System.Diagnostics;
-using osu.Framework.Input;
 using osu.Framework.MathUtils;
 using OpenTK;
 using OpenTK.Graphics;
 using osu.Framework.Graphics.Shapes;
+using osu.Framework.Input.EventArgs;
 using osu.Framework.Input.States;
 using OpenTK.Input;
