+++ osu.Framework/Graphics/Cursor/ContextMenuContainer.cs
@@ -7,6 +7,7 @@ using OpenTK.Input;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.UserInterface;
 using osu.Framework.Input;
+using osu.Framework.Input.EventArgs;
 using osu.Framework.Input.States;
 
 namespace osu.Framework.Graphics.Cursor
