+++ osu.Framework/Graphics/Drawable.cs
@@ -26,9 +26,10 @@ using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using osu.Framework.Development;
+using osu.Framework.Input.EventArgs;
 using osu.Framework.Input.States;
 using osu.Framework.MathUtils;
-using JoystickEventArgs = osu.Framework.Input.JoystickEventArgs;
+using JoystickEventArgs = osu.Framework.Input.EventArgs.JoystickEventArgs;
 
 namespace osu.Framework.Graphics
 {
