+++ osu.Framework/Graphics/Visualisation/VisualisedDrawable.cs
@@ -4,7 +4,6 @@
 using System;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Sprites;
-using osu.Framework.Input;
 using OpenTK;
 using OpenTK.Graphics;
 using osu.Framework.Extensions.Color4Extensions;
@@ -12,6 +11,7 @@ using OpenTK.Input;
 using osu.Framework.Graphics.Shapes;
 using System.Collections.Generic;
 using osu.Framework.Extensions.IEnumerableExtensions;
+using osu.Framework.Input.EventArgs;
 using osu.Framework.Input.States;
 
 namespace osu.Framework.Graphics.Visualisation
