+++ osu.Framework/Input/Bindings/KeyBindingContainer.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using osu.Framework.Extensions.IEnumerableExtensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
+using osu.Framework.Input.EventArgs;
 using osu.Framework.Input.States;
 using osu.Framework.Logging;
 using OpenTK;
