+++ osu.Framework/Input/EventArgs/JoystickEventArgs.cs
@@ -1,9 +1,10 @@
 ﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
-namespace osu.Framework.Input
+namespace osu.Framework.Input.EventArgs
 {
-    public class MouseUpEventArgs : MouseEventArgs
+    public class JoystickEventArgs : System.EventArgs
     {
+        public JoystickButton Button;
     }
 }
