+++ osu.Framework/Input/EventArgs/MouseDownEventArgs.cs
@@ -1,7 +1,7 @@
 // Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
-namespace osu.Framework.Input
+namespace osu.Framework.Input.EventArgs
 {
     public class MouseDownEventArgs : MouseEventArgs
     {
