+++ osu.Framework/Input/EventArgs/MouseEventArgs.cs
@@ -2,12 +2,11 @@
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
 using OpenTK.Input;
-using System;
 
-namespace osu.Framework.Input
+namespace osu.Framework.Input.EventArgs
 {
-    public class KeyUpEventArgs : EventArgs
+    public class MouseEventArgs : System.EventArgs
     {
-        public Key Key;
+        public MouseButton Button;
     }
 }
