+++ osu.Framework/Input/EventArgs/MouseUpEventArgs.cs
@@ -1,13 +1,9 @@
 ﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
-using System;
-using OpenTK.Input;
-
-namespace osu.Framework.Input
+namespace osu.Framework.Input.EventArgs
 {
-    public class MouseEventArgs : EventArgs
+    public class MouseUpEventArgs : MouseEventArgs
     {
-        public MouseButton Button;
     }
 }
