+++ osu.Framework/Input/PassThroughInputManager.cs
@@ -4,10 +4,12 @@
 using System.Collections.Generic;
 using System.Linq;
 using osu.Framework.Graphics;
+using osu.Framework.Input.EventArgs;
 using osu.Framework.Input.StateChanges;
 using osu.Framework.Input.States;
 using OpenTK;
 using OpenTK.Input;
+using JoystickEventArgs = osu.Framework.Input.EventArgs.JoystickEventArgs;
 
 namespace osu.Framework.Input
 {
