+++ osu.Framework/Input/States/InputState.cs
@@ -1,11 +1,9 @@
 // Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
-using System;
-
 namespace osu.Framework.Input.States
 {
-    public class InputState : EventArgs
+    public class InputState : System.EventArgs
     {
         public IKeyboardState Keyboard;
         public IMouseState Mouse;
